Fix flakiness in test
Given the database is already full of data, selecting the first 2 messages from the GET all messages could select any old pair of messages. The order is undefined from the database but, by chance, is often ordered with the most recent first making the test appear to be correct when it was first written.